check for unused packages after install command
