+++ lib/commands/install.js
@@ -13,6 +13,7 @@ var semver = require('semver'),
     argParse = require('../args').parse,
     tar = require('../tar'),
     tree = require('../tree'),
+    clean = require('./clean'),
     async = require('async'),
     path = require('path'),
     fs = require('fs');
@@ -239,7 +240,13 @@ exports.installTree = function (packages, opt, callback) {
         else {
             process.nextTick(cb);
         }
-    }, callback);
+    }, function (err) {
+        if (err) {
+            return callback(err);
+        }
+        // report packages that are no-longer used
+        exports.checkUnused(packages, opt, callback);
+    });
 };
 
 
@@ -532,5 +539,36 @@ exports.installURL = function (url, opt, callback) {
             return callback(err);
         }
         exports.installFile(filename, opt, callback);
+    });
+};
+
+
+/**
+ * Checks the packages directory for packages not in the provided version
+ * tree.
+ *
+ * @param {Object} packages - the version tree to compare against
+ * @param {Object} opt - the options object
+ * @param {Function} callback
+ */
+
+exports.checkUnused = function (packages, opt, callback) {
+    clean.unusedDirsTree(packages, opt, function (err, dirs) {
+        if (err) {
+            return callback(err);
+        }
+        if (dirs.length) {
+            var names = dirs.map(function (d) {
+                return path.relative('.', d);
+            });
+            console.log(
+                '\n' +
+                'The following packages are no longer required, ' +
+                'and can be removed\nby running "kanso clean":\n' +
+                '    ' + names.join('\n    ') +
+                '\n'
+            );
+        }
+        return callback();
     });
 };
