+++ lib/commands/update.js
@@ -4,7 +4,6 @@
 
 var path = require('path'),
     semver = require('semver'),
-    clean = require('./clean'),
     install = require('./install'),
     tree = require('../tree');
     utils = require('../utils'),
@@ -85,41 +84,10 @@ exports.update = function (deps, opt, callback) {
                     if (err) {
                         return callback(err);
                     }
-                    exports.checkUnused(updated, opt, callback);
+                    install.checkUnused(updated, opt, callback);
                 });
             }
         );
-    });
-};
-
-
-/**
- * Checks the packages directory for packages not in the provided version
- * tree.
- *
- * @param {Object} packages - the version tree to compare against
- * @param {Object} opt - the options object
- * @param {Function} callback
- */
-
-exports.checkUnused = function (packages, opt, callback) {
-    clean.unusedDirsTree(packages, opt, function (err, dirs) {
-        if (err) {
-            return callback(err);
-        }
-        if (dirs.length) {
-            var names = dirs.map(function (d) {
-                return path.relative('.', d);
-            });
-            console.log(
-                '\n' +
-                'The following packages are no longer required, ' +
-                'and can be removed\nby running "kanso clean":\n' +
-                '    ' + names.join('\n    ') +
-                '\n'
-            );
-        }
-        return callback();
     });
 };
